비밀번호 변경 API, 이벤트 상세페이지 API 연결 완료

+++ app/src/main/java/com/manu/todayhouse/src/main/home/popluar/eventpage/model/Result.kt
@@ -3,5 +3,5 @@ package com.manu.todayhouse.src.main.home.popluar.eventpage.model
 data class Result(
     val bannerPic: String,
     val due: String,
-    val eventId: Int
+    val eventId: Long
 )
