+++ app/src/main/java/com/manu/todayhouse/src/main/store/storehome/model/TodaysDeal.kt
@@ -1,16 +1,16 @@
 package com.manu.todayhouse.src.main.store.storehome.model
 
 data class TodaysDeal(
-    val companyId: Int,
+    val companyId: Long,
     val companyName: String,
     val due: String,
     val hotDealThumbnail: String,
-    val itemId: Int,
+    val itemId: Long,
     val itemName: String,
     val price: String,
     val reviewNum: Int,
     val saleRate: String,
     val score: Double,
     val subCategory: String,
-    val subCategoryId: Int
+    val subCategoryId: Long
 )
